feat: 검색 결과 피드 연결
- 검색 결과 피드 연결, 해당 피드로 포커스
- 홈 태그 결과 피드 연결

+++ app/src/main/java/com/codepatissier/keki/src/main/consumer/store/storefeed/model/Feed.kt
@@ -5,7 +5,7 @@ data class Feed(
     val description: String,
     val dessertName: String,
     val like: Boolean,
-    val postIdx: Int,
+    val postIdx: Long,
     val postImgUrls: List<String>,
     val storeProfileImg: String,
     val tags: List<String>
